REF: Split Birdeye schema.py into a domain sub-package
The single `schema.py` had grown to ~78 KB after the Token Overview
rewrite and was set to keep growing significantly as we wrap the new
v3 token endpoints, holder/transfer/trending endpoints and the multi-
variant flavours of each. Reading through one monolithic file becomes
impractical at that scale.

Split `cyhole/birdeye/schema.py` into a `schema/` sub-package, one file
per logical API domain:

- `token_list.py`        Token - List (V1) classes
- `security.py`          Token - Security classes
- `token_stats.py`       Token - Overview classes
- `creation_trending.py` Token - Creation Info classes (Trending will
                         join later)
- `price.py`             Price, Price - Multiple, Price - Historical,
                         Price Volume - Single and Multi classes
- `transactions.py`      Trades - Token, Trades - Pair, OHLCV variants
- `wallet.py`            Wallet - Supported Networks class

`schema/__init__.py` re-exports every public name, so existing imports
like `from cyhole.birdeye.schema import GetTokenListResponse` keep
working without any client/test/docs change.

Adopt this layout as the project convention for any Interaction whose
schema file grows too large to navigate.
