+++ src/cyhole/birdeye/schema/token_stats.py
@@ -1,164 +1,6 @@
-from typing import Any
 from datetime import datetime
 
 from pydantic import BaseModel, Field, field_validator
-
-# classes used on GET "Token List" endpoint
-class GetTokenListInfo(BaseModel):
-    name: str | None                = None
-    symbol: str | None              = None
-    price: float
-    address: str
-    decimals: int
-    liquidity: float
-    volume_24h_usd: float           = Field(alias = "v24hUSD")
-    market_cap: float               = Field(alias = "mc")
-    volume_24h_change: float | None = Field(alias = "v24hChangePercent", default = None)
-    last_trade_unix_time: float     = Field(alias = "lastTradeUnixTime")
-    logo_uri: str  | None           = Field(alias = "logoURI", default = None)
-
-class GetTokenListData(BaseModel):
-    total: int | None = None
-    update_time: datetime = Field(alias = "updateTime")
-    update_unix_time: int = Field(alias = "updateUnixTime")
-    tokens: list[GetTokenListInfo]
-
-    @field_validator("update_time")
-    def parse_update_time(cls, update_time_raw: str | datetime) -> datetime:
-        if isinstance(update_time_raw, str):
-            return datetime.strptime(update_time_raw, "%Y-%m-%dT%H:%M:%S")
-        return update_time_raw
-
-class GetTokenListResponse(BaseModel):
-    """
-        Model used to represent the **Token - List** endpoint from birdeye API.
-    """
-    data: GetTokenListData
-    success: bool
-
-# classes used on GET "Token - Security" endpoint
-class GetTokenSecurityDataSolana(BaseModel):
-    """
-        Solana-specific security/risk profile of a token returned by the Birdeye
-        **Token - Security** endpoint.
-
-        All ownership and authority fields are optional because Birdeye returns
-        `null` for them when the token has renounced/cleared the corresponding
-        on-chain authority. Balance and percentage fields refer to the holdings
-        of the related authority over the token's total supply.
-
-        Attributes:
-            creator_address: address that created the token (alias `creatorAddress`); `None` if unknown.
-            creator_owner_address: owner of `creator_address` at the time of token creation
-                (alias `creatorOwnerAddress`); `None` if the creator was an EOA or no longer tracked.
-            owner_address: current on-chain owner of the token authority (alias `ownerAddress`);
-                `None` when ownership has been renounced.
-            owner_of_owner_address: owner of `owner_address` (alias `ownerOfOwnerAddress`);
-                `None` if owner is an EOA or ownership has been renounced.
-            creation_transaction: signature of the transaction that created the token
-                (alias `creationTx`); `None` if unknown.
-            creation_time_unix: unix-second timestamp of the creation transaction (alias `creationTime`).
-            creation_slot: slot of the creation transaction (alias `creationSlot`).
-            mint_transaction: signature of the mint transaction (alias `mintTx`).
-            mint_time_unix: unix-second timestamp of the mint transaction (alias `mintTime`).
-            mint_slot: slot of the mint transaction (alias `mintSlot`).
-            mintable: whether the token can still be minted; `None` if the API cannot determine it.
-            renounce: whether the mint authority has been renounced; `None` if undetermined.
-            creator_balance: token balance currently held by the creator address (alias `creatorBalance`).
-            creator_percentage: creator balance expressed as a fraction (0..1) of total supply
-                (alias `creatorPercentage`).
-            owner_balance: token balance currently held by the owner address (alias `ownerBalance`).
-            owner_percentage: owner balance expressed as a fraction (0..1) of total supply
-                (alias `ownerPercentage`).
-            metaplex_update_authority: Metaplex metadata update authority (alias `metaplexUpdateAuthority`).
-            metaplex_owner_update_authority: owner of the Metaplex update authority address
-                (alias `metaplexOwnerUpdateAuthority`); `None` when no upstream owner is tracked.
-            metaplex_update_balance: token balance held by the Metaplex update authority
-                (alias `metaplexUpdateAuthorityBalance`).
-            metaplex_update_percentage: Metaplex update authority balance expressed as a fraction (0..1)
-                of total supply (alias `metaplexUpdateAuthorityPercent`).
-            mutable_metadata: whether the on-chain Metaplex metadata of the token is still mutable
-                (alias `mutableMetadata`).
-            top_10_holder_balance: cumulative token balance of the top-10 holders (alias `top10HolderBalance`).
-            top_10_holder_percentage: top-10 holders' cumulative balance as a fraction (0..1) of total
-                supply (alias `top10HolderPercent`).
-            top_10_user_balance: cumulative token balance of the top-10 *user* wallets (excluding programs)
-                (alias `top10UserBalance`).
-            top_10_user_percentage: top-10 users' cumulative balance as a fraction (0..1) of total supply
-                (alias `top10UserPercent`).
-            is_true_token: Birdeye flag marking the token as one of its verified entries
-                (alias `isTrueToken`); `None` when unverified.
-            fake_token: Birdeye flag marking the token as a known scam/fake (alias `fakeToken`);
-                `None` when Birdeye has not categorised it.
-            total_supply: total on-chain supply of the token, expressed in UI units (alias `totalSupply`).
-            pre_market_holder: list of addresses that received tokens before any public market existed
-                (alias `preMarketHolder`).
-            lock_info: free-form payload describing on-chain locks (e.g. team vesting); `None` when no
-                lock information is known (alias `lockInfo`).
-            freezeable: whether the SPL freeze authority can still freeze accounts; `None` if undetermined.
-            freeze_authority: address that owns the freeze authority (alias `freezeAuthority`);
-                `None` when the freeze authority has been renounced.
-            transfer_fee_enable: whether the Token-2022 transfer-fee extension is active
-                (alias `transferFeeEnable`); `None` if the extension is absent or undetermined.
-            transfer_fee_data: free-form payload describing the active transfer-fee parameters
-                (alias `transferFeeData`); `None` when no fee data is available.
-            is_token_2022: whether the token is an SPL Token-2022 mint (alias `isToken2022`).
-            non_transferable: whether the Token-2022 non-transferable extension is active
-                (alias `nonTransferable`); `None` if undetermined.
-            jup_strict_list: whether the token is part of Jupiter's strict token list
-                (alias `jupStrictList`); `None` when Birdeye did not populate the flag.
-    """
-    creator_address: str | None = Field(alias = "creatorAddress", default = None)
-    creator_owner_address: str | None = Field(alias = "creatorOwnerAddress", default = None)
-    owner_address: str | None = Field(alias = "ownerAddress", default = None)
-    owner_of_owner_address: str | None = Field(alias = "ownerOfOwnerAddress", default = None)
-    creation_transaction: str | None = Field(alias = "creationTx", default = None)
-    creation_time_unix: int | None = Field(alias = "creationTime", default = None)
-    creation_slot: int | None = Field(alias = "creationSlot", default = None)
-    mint_transaction: str | None = Field(alias = "mintTx", default = None)
-    mint_time_unix: int | None = Field(alias = "mintTime", default = None)
-    mint_slot: int | None = Field(alias = "mintSlot", default = None)
-    mintable: bool | None = None
-    renounce: bool | None = None
-    creator_balance: float | None = Field(alias = "creatorBalance", default = None)
-    creator_percentage: float | None = Field(alias = "creatorPercentage", default = None)
-    owner_balance: float | None = Field(alias = "ownerBalance", default = None)
-    owner_percentage: float | None = Field(alias = "ownerPercentage", default = None)
-    metaplex_update_authority: str = Field(alias = "metaplexUpdateAuthority")
-    metaplex_owner_update_authority: str | None = Field(alias = "metaplexOwnerUpdateAuthority", default = None)
-    metaplex_update_balance: float = Field(alias = "metaplexUpdateAuthorityBalance")
-    metaplex_update_percentage: float = Field(alias = "metaplexUpdateAuthorityPercent")
-    mutable_metadata: bool = Field(alias = "mutableMetadata")
-    top_10_holder_balance: float = Field(alias = "top10HolderBalance")
-    top_10_holder_percentage: float = Field(alias = "top10HolderPercent")
-    top_10_user_balance: float = Field(alias = "top10UserBalance")
-    top_10_user_percentage: float = Field(alias = "top10UserPercent")
-    is_true_token: bool | None = Field(alias = "isTrueToken", default = None)
-    fake_token: bool | None = Field(alias = "fakeToken", default = None)
-    total_supply: float = Field(alias = "totalSupply")
-    pre_market_holder: list[str] = Field(alias = "preMarketHolder")
-    lock_info: Any | None = Field(alias = "lockInfo", default = None)
-    freezeable: bool | None = None
-    freeze_authority: str | None = Field(alias = "freezeAuthority", default = None)
-    transfer_fee_enable: bool | None = Field(alias = "transferFeeEnable", default = None)
-    transfer_fee_data: Any | None = Field(alias = "transferFeeData", default = None)
-    is_token_2022: bool = Field(alias = "isToken2022")
-    non_transferable: bool | None = Field(alias = "nonTransferable", default = None)
-    jup_strict_list: bool | None = Field(alias = "jupStrictList", default = None)
-
-class GetTokenSecurityResponse(BaseModel):
-    """
-        Model used to represent the **Token - Security** endpoint from birdeye API.
-
-        Attributes:
-            data: security profile of the token. When the request targets Solana the payload is
-                parsed into the typed [`GetTokenSecurityDataSolana`][cyhole.birdeye.schema.GetTokenSecurityDataSolana]
-                schema; for every other supported chain the payload is left as a free-form
-                dictionary because Birdeye returns a chain-specific (EVM-style) shape.
-            success: `True` when the API call completed without errors.
-    """
-    data: GetTokenSecurityDataSolana | dict[str, Any]
-    success: bool
 
 # classes used on GET "Token - Overview" endpoint
 class GetTokenOverviewData(BaseModel):
@@ -702,269 +544,3 @@ class GetTokenOverviewResponse(BaseModel):
     data: GetTokenOverviewData
     success: bool
 
-# classes used on GET "Token - Creation Token Info" endpoint
-class GetTokenCreationInfoData(BaseModel):
-    """
-        Creation metadata of a token returned by the Birdeye
-        **Token - Creation Token Info** endpoint.
-
-        Attributes:
-            transaction_hash: hash of the on-chain transaction that minted the token
-                (alias `txHash`).
-            slot: slot in which the mint transaction was processed (Solana-style ordinal).
-            token_address: contract address of the freshly created token (alias `tokenAddress`).
-            decimals: number of decimal places used by the token.
-            owner: address that signed and paid for the creation transaction.
-            block_unix_time: timestamp of the block containing the creation transaction,
-                expressed as unix seconds (alias `blockUnixTime`).
-            block_human_time: human-readable ISO-8601 timestamp of the same block as
-                `block_unix_time` (alias `blockHumanTime`).
-    """
-    transaction_hash: str = Field(alias = "txHash")
-    slot: int
-    token_address: str = Field(alias = "tokenAddress")
-    decimals: int
-    owner: str
-    block_unix_time: int = Field(alias = "blockUnixTime")
-    block_human_time: datetime = Field(alias = "blockHumanTime")
-
-    @field_validator("block_human_time")
-    def parse_block_human_time(cls, dt_raw: str | datetime) -> datetime:
-        if isinstance(dt_raw, str):
-            # Birdeye returns ISO timestamps with a trailing "Z" and optional milliseconds.
-            return datetime.fromisoformat(dt_raw.replace("Z", "+00:00"))
-        return dt_raw
-
-class GetTokenCreationInfoResponse(BaseModel):
-    """
-        Model used to represent the **Token - Creation Token Info** endpoint from birdeye API.
-
-        Attributes:
-            data: token creation metadata payload.
-            success: `True` when the API call completed without errors.
-    """
-    data: GetTokenCreationInfoData
-    success: bool
-
-
-# classes used on GET "Price" endpoint
-class GetPriceData(BaseModel):
-    value: float
-    liquidity: float | None = None
-    update_human_time: datetime = Field(alias = "updateHumanTime")
-    update_unix_time: int = Field(alias = "updateUnixTime")
-
-    @field_validator("update_human_time")
-    def parse_update_human_time(cls, update_human_time_raw: str | datetime) -> datetime:
-        if isinstance(update_human_time_raw, str):
-            return datetime.strptime(update_human_time_raw, "%Y-%m-%dT%H:%M:%S")
-        return update_human_time_raw
-
-class GetPriceResponse(BaseModel):
-    """
-        Model used to represent the **Price** endpoint from birdeye API.
-    """
-    data: GetPriceData
-    success: bool
-
-# classes used on GET "Price - Multiple" endpoint
-class GetPriceMultipleData(GetPriceData):
-    price_change_24h: float = Field(alias = "priceChange24h")
-
-class GetPriceMultipleResponse(BaseModel):
-    """
-        Model used to represent the **Price - Multiple** endpoint from birdeye API.
-    """
-    data: dict[str, GetPriceMultipleData]
-    success: bool
-
-# classes used on GET "Price - Historical" endpoint
-class GetPriceHistoricalMeasure(BaseModel):
-    unix_time: int = Field(alias = "unixTime")
-    value: float
-
-class GetPriceHistoricalData(BaseModel):
-    items: list[GetPriceHistoricalMeasure]
-
-class GetPriceHistoricalResponse(BaseModel):
-    """
-        Model used to represent the **Price - Historical** endpoint from birdeye API.
-    """
-    data: GetPriceHistoricalData
-    success: bool
-
-# classes used on GET "Price Volume - Single Token" endpoint
-class GetPriceVolumeSingleData(BaseModel):
-    price: float
-    update_unix_time: int = Field(alias = "updateUnixTime")
-    update_human_time: datetime = Field(alias = "updateHumanTime")
-    volume_usd: float = Field(alias = "volumeUSD")
-    volume_change_percent: float = Field(alias = "volumeChangePercent")
-    price_change_percent: float = Field(alias = "priceChangePercent")
-
-    @field_validator("update_human_time")
-    def parse_update_human_time(cls, update_human_time_raw: str | datetime) -> datetime:
-        if isinstance(update_human_time_raw, str):
-            return datetime.strptime(update_human_time_raw, "%Y-%m-%dT%H:%M:%S")
-        return update_human_time_raw
-
-class GetPriceVolumeSingleResponse(BaseModel):
-    """
-        Model used to represent the **Price Volume - Single Token** endpoint from birdeye API.
-    """
-    data: GetPriceVolumeSingleData
-    success: bool
-
-# classes used on POST "Price Volume - Multiple Token" endpoint
-class PostPriceVolumeMultiData(GetPriceVolumeSingleData):
-    address: str
-
-class PostPriceVolumeMultiResponse(BaseModel):
-    """
-        Model used to represent the **Price Volume - Multiple Token** endpoint from birdeye API.
-    """
-    data: list[PostPriceVolumeMultiData]
-    success: bool
-
-# classes used on GET "Trades - Token" endpoint
-class GetTradesTokenTradeToken(BaseModel):
-    symbol: str
-    decimals: int
-    address: str
-    amount: int
-    type: str
-    type_swap: str = Field(alias = "typeSwap")
-    ui_amount: float = Field(alias = "uiAmount")
-    price: float | None = None
-    nearest_price: float = Field(alias = "nearestPrice")
-    change_amount: float = Field(alias = "changeAmount")
-    ui_change_amount: float = Field(alias = "uiChangeAmount")
-    icon: str | None = None
-
-class GetTradesTokenTrade(BaseModel):
-    volume: float
-    volume_usd: float = Field(alias = "volumeUSD")
-    trade_hash: str = Field(alias = "txHash")
-    slot: int
-    source: str
-    block_unix_time: int = Field(alias = "blockUnixTime")
-    block_human_time: datetime = Field(alias = "blockHumanTime")
-    trade_type: str = Field(alias = "txType")
-    address: str
-    owner: str
-    trade_status: str = Field(alias = "txStatus")
-    outliers: bool
-    nearest_price_base_coin: float = Field(alias = "nearestPriceBaseCoin")
-    nearest_price_quote_coin: float = Field(alias = "nearestPriceQuoteCoin")
-    ins_index: int = Field(alias = "insIndex")
-    inner_ins_index: int | None = Field(alias = "innerInsIndex", default = None)
-    be_bevenue: str | None = Field(alias = "beRevenue", default = None)
-    event_type: str = Field(alias = "eventType")
-    side: str
-    price_pair: float = Field(alias = "pricePair")
-    alias: str | None = None
-    platform: str
-    toke_price: float = Field(alias = "tokenPrice")
-    proce_mark: bool = Field(alias = "priceMark")
-    network: str
-    trade_from: GetTradesTokenTradeToken = Field(alias = "from")
-    trade_to: GetTradesTokenTradeToken = Field(alias = "to")
-
-    @field_validator("block_human_time")
-    def parse_update_human_time(cls, dt_raw: str | datetime) -> datetime:
-        if isinstance(dt_raw, str):
-            return datetime.strptime(dt_raw, "%Y-%m-%dT%H:%M:%S")
-        return dt_raw
-
-class GetTradesTokenData(BaseModel):
-    items: list[GetTradesTokenTrade]
-    has_next: bool = Field(alias = "hasNext")
-
-class GetTradesTokenResponse(BaseModel):
-    """
-        Model used to represent the **Trades - Token** endpoint from birdeye API.
-    """
-    data: GetTradesTokenData
-    success: bool
-
-# classes used on GET "Trades - Pair" endpoint
-class GetTradesPairTradeToken(BaseModel):
-    symbol: str
-    decimals: int
-    address: str
-    amount: int
-    type: str
-    type_swap: str = Field(alias = "typeSwap")
-    ui_amount: float = Field(alias = "uiAmount")
-    price: float | None = None
-    nearest_price: float = Field(alias = "nearestPrice")
-    change_amount: float = Field(alias = "changeAmount")
-    ui_change_amount: float = Field(alias = "uiChangeAmount")
-
-class GetTradesPairTrade(BaseModel):
-    trade_hash: str = Field(alias = "txHash")
-    source: str
-    block_unix_time: int = Field(alias = "blockUnixTime")
-    address: str
-    owner: str
-    trade_from: GetTradesPairTradeToken = Field(alias = "from")
-    trade_to: GetTradesPairTradeToken = Field(alias = "to")
-
-class GetTradesPairData(BaseModel):
-    items: list[GetTradesPairTrade]
-    has_next: bool = Field(alias = "hasNext")
-
-class GetTradesPairResponse(BaseModel):
-    """
-        Model used to represent the **Trades - Pair** endpoint from birdeye API.
-    """
-    data: GetTradesPairData
-    success: bool
-
-# classes used on GET "OHLCV - Token/Pair" endpoint
-class GetOHLCVInterval(BaseModel):
-    close: float = Field(alias = "c")
-    high: float  = Field(alias = "h")
-    low: float  = Field(alias = "l")
-    open: float  = Field(alias = "o")
-    type: str
-    unix_time: int  = Field(alias = "unixTime")
-
-class GetOHLCVTokenPairInterval(GetOHLCVInterval):
-    address: str
-    volume: float  = Field(alias = "v")
-
-class GetOHLCVTokenPairData(BaseModel):
-    items: list[GetOHLCVTokenPairInterval]
-
-class GetOHLCVTokenPairResponse(BaseModel):
-    """
-        Model used to represent the **OHLCV - Token/Pair** endpoint from birdeye API.
-    """
-    data: GetOHLCVTokenPairData
-    success: bool
-
-# classes used on GET "OHLCV - Base/Quote" endpoint
-class GetOHLCVBaseQuoteInterval(GetOHLCVInterval):
-    base_address: str = Field(alias = "baseAddress")
-    quote_address: str = Field(alias = "quoteAddress")
-    base_volume: float = Field(alias = "vBase")
-    quote_volume: float = Field(alias = "vQuote")
-
-class GetOHLCVBaseQuoteData(BaseModel):
-    items: list[GetOHLCVBaseQuoteInterval]
-
-class GetOHLCVBaseQuoteResponse(BaseModel):
-    """
-        Model used to represent the **OHLCV - Token/Pair** endpoint from birdeye API.
-    """
-    data: GetOHLCVBaseQuoteData
-    success: bool
-
-# classes used on GET "Wallet - Supported Networks" endpoint
-class GetWalletSupportedNetworksResponse(BaseModel):
-    """
-        Model used to represent the **Wallet - Supported Networks** endpoint from birdeye API.
-    """
-    data: list[str]
-    success: bool
