+++ src/cyhole/birdeye/schema/wallet.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+# classes used on GET "Wallet - Supported Networks" endpoint
+class GetWalletSupportedNetworksResponse(BaseModel):
+    """
+        Model used to represent the **Wallet - Supported Networks** endpoint from birdeye API.
+    """
+    data: list[str]
+    success: bool
